Download datasets via the Data Retriever.

Directly downloads data files with no processing, allowing downloading of non-tabular data.

download(
  dataset,
  path = "./",
  quiet = FALSE,
  sub_dir = "",
  debug = FALSE,
  use_cache = TRUE
)

Arguments

dataset
the name of the dataset that you wish to download

path
the path where the data should be downloaded to

quiet
logical, if true retriever runs in quiet mode

sub_dir
downloaded dataset is stored into a custom subdirectory.

debug
Setting TRUE helps in debugging in case of errors

use_cache
Setting FALSE reinstalls scripts even if they are already installed

Examples

# \donttest{
rdataretriever::download("plant-comp-ok")

#> Error: Python module retriever was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python3
#> libpython:      /usr/lib/python3.8/config-3.8-x86_64-linux-gnu/libpython3.8.so
#> pythonhome:     //usr://usr
#> version:        3.8.2 (default, Jul 16 2020, 14:00:26)  [GCC 9.3.0]
#> numpy:          /usr/lib/python3/dist-packages/numpy
#> numpy_version:  1.17.4
#> retriever:      [NOT FOUND]
#> 

# downloaded files will be copied to your working directory
# when no path is specified
dir()

#> [1] "check_for_updates.html"            "check_retriever_availability.html"
#> [3] "commit_log.html"                   "commit.html"                      
#> [5] "data_retriever_version.html"       "dataset_names.html"               
#> [7] "datasets.html"                     "index.html"                       
#> [9] "Rplot001.png"                     

# }